refactor(region-map): extract handleRegionAction into new module and wire into region_map_runtime.js
